+++ mobile/apps/locker/lib/ui/drawer/settings_widgets.dart
@@ -6,6 +6,7 @@ import "package:flutter/material.dart";
 import "package:locker/l10n/l10n.dart";
 import "package:locker/ui/drawer/about_section_widget.dart";
 import "package:locker/ui/drawer/account_section_widget.dart";
+import "package:locker/ui/drawer/general_section_widget.dart";
 import "package:locker/ui/drawer/security_section_widget.dart";
 import "package:locker/ui/drawer/social_section_widget.dart";
 import "package:locker/ui/drawer/support_section_widget.dart";
@@ -54,6 +55,11 @@ class SettingsWidgets extends StatelessWidget {
         ]);
       }
     }
+
+    contents.addAll([
+      const GeneralSectionWidget(),
+      sectionSpacing,
+    ]);
 
     contents.addAll([
       const SupportSectionWidget(),