+++ mobile/apps/photos/lib/ui/notification/update/change_log_page.dart
@@ -103,16 +103,6 @@ class _ChangeLogPageState extends State<ChangeLogPage> {
         context.l10n.cLDesc2,
         isFeature: true,
       ),
-      ChangeLogEntry(
-        context.l10n.cLTitle3,
-        context.l10n.cLDesc3,
-        isFeature: true,
-      ),
-      ChangeLogEntry(
-        context.l10n.cLTitle4,
-        context.l10n.cLDesc4,
-        isFeature: false,
-      ),
     ]);
     return Container(
       padding: const EdgeInsets.only(left: 16),