+++ server/pkg/controller/collections/share.go
@@ -146,6 +146,9 @@ func (c *CollectionController) UnShare(ctx *gin.Context, cID int64, fromUserID i
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "")
 	}
+	if err := c.removeUserSocialActivity(ctx.Request.Context(), cID, toUserID); err != nil {
+		return nil, err
+	}
 	err = c.CastRepo.RevokeForGivenUserAndCollection(ctx, cID, toUserID)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "")
@@ -181,6 +184,9 @@ func (c *CollectionController) Leave(ctx *gin.Context, cID int64) error {
 	err = c.CollectionRepo.UnShare(cID, userID)
 	if err != nil {
 		return stacktrace.Propagate(err, "")
+	}
+	if err := c.removeUserSocialActivity(ctx.Request.Context(), cID, userID); err != nil {
+		return err
 	}
 	return nil
 }